budget-todo-fast-api-be/requirements.txt
fastapi
uvicorn
sqlalchemy
# mysqlclient
databases
pymysql
mysql-connector-python
python-keycloak
pydantic-settings
python-dotenv
asgiref
cffi
click
cryptography
h11
pycparser
pydantic
PyJWT
starlette
typing-extensions